ai-service/requirements.txt
# SLIB AI Service - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# HTTP Client
httpx>=0.26.0

# Configuration
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Utilities
python-dotenv>=1.0.0

# LangChain & AI
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-ollama>=0.0.2

# Database
sqlalchemy>=2.0.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
pgvector>=0.2.4
asyncpg>=0.29.0

# Document Processing
python-docx>=1.1.0
PyPDF2>=3.0.0
python-multipart>=0.0.6

# Optional: For production
# gunicorn==21.2.0
# redis==5.0.1

